Fixed bugs in validation_step. NSDR is now correct.
The setting of dataloader for validation set is wrong. It is supposed to load 5 sources (mix, drums, bass, other, vocal). Instead it loaded only 4 sources (drums, bass, other, vocals)

Due to the fact that only 4 sources were loaded, line 529 was actually the audio for drums. Then the model tried to use the drum audio to predict the 4 sources. That was the reason why the NSDR for drum was super high, but the NSDRs for bass, other, and vocals were super low.

This commit also removed unecessary debugging codes.